fixing pin input
Testing digital input for pin 19 (and any pin which is part of port 2) lead to the discovery of out-of-bounds behaviour in the poll method's DIGITAL_MESSAGE branch. Removing this error is possible as protection against setting non-existing pins exists in the user-facing methods. Here, we assume that the attached board from where the messages are coming knows what pins exist and as such we can simply ignore the attemped access of high pins. (the code handing the distribute of port values across pins might need to be redone to account for situations where ports don't actually attach to 8 pins)