fixing pin mode setting
for some reason, in the past when a user would set the mode of a pin it would actually set the list of all possible modes for that pin to only contain the selected mode. It didn't actually set the pin mode. It set it on the board though, so in the case of pin input, when a message came from the board it would be ignored for that pin as the mode was incorreect.

I've corrected this, and added extra protection on the methods.