send REPORT_FIRMWARE message when connection is established
Previously, it was assumed that establishing a connection to a device would cause it to start, but after testing this has proven not to be the case on all systems (namely; Windows 10)

After establishing a connection, the client expects to receive the device's firmware name and version as a signal to show that the device is ready and that the rest of the setup procedure can continue. This information is sent automatically when the device starts. If the device does not restart, then the client wil wait forever. The changes here mean that a request for the information is sent after the connection is established.